refactor: DB 관련 파일 패키지 분리
