+++ app/src/main/java/woowacourse/shopping/data/ShoppingCartDatabase.kt
@@ -4,6 +4,9 @@ import android.content.Context
 import androidx.room.Database
 import androidx.room.Room
 import androidx.room.RoomDatabase
+import woowacourse.shopping.data.product.ProductEntity
+import woowacourse.shopping.data.shoppingcart.ShoppingCartDao
+import woowacourse.shopping.data.shoppingcart.ShoppingCartEntity
 
 @Database(entities = [ProductEntity::class, ShoppingCartEntity::class], version = 2)
 abstract class ShoppingCartDatabase : RoomDatabase() {
