+++ app/src/main/java/woowacourse/shopping/data/product/ProductRepository.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data
+package woowacourse.shopping.data.product
 
 import woowacourse.shopping.domain.Product
